fix: use consistent %ld/static_cast<long> for all timing log format specifiers in gateway.cpp
Agent-Logs-Url: https://github.com/rdkcentral/firebolt-cpp-transport/sessions/b93c5ecf-14e3-47e9-8a25-e8a6660afffa

+++ src/gateway.cpp
@@ -534,10 +534,10 @@ public:
         FIREBOLT_LOG_INFO("Gateway", "[subscribe] waiting for subscribe ACK for '%s'...", event.c_str());
         auto t0_sub = std::chrono::steady_clock::now();
         auto result = request(event, params, id).get();
-        FIREBOLT_LOG_INFO("Gateway", "[subscribe] ACK for '%s' received in %lld ms", event.c_str(),
-                          static_cast<long long>(
-                              std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::steady_clock::now() - t0_sub)
-                                  .count()));
+        FIREBOLT_LOG_INFO("Gateway", "[subscribe] ACK for '%s' received in %ld ms", event.c_str(),
+                          static_cast<long>(std::chrono::duration_cast<std::chrono::milliseconds>(
+                                                std::chrono::steady_clock::now() - t0_sub)
+                                                .count()));
 
         if (!result)
         {
